harmonize error definitions
