+++ rust/flight-fusion-ipc/src/error.rs
